feat(common): move shared exceptions and constants to support:common module
