--- coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/error/ErrorCode.java
@@ -1,6 +1,0 @@
-package dev.be.coupon.api.common.support.error;
-
-public enum ErrorCode {
-    E500,
-    E400, E401, E403, E404
-}
