--- coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/filter/v1/HttpLogMessage.java
@@ -1,14 +1,0 @@
-package dev.be.coupon.api.common.support.filter.v1;
-
-import org.springframework.http.HttpStatus;
-
-public record HttpLogMessage(
-        String httpMethod,
-        String url,
-        HttpStatus httpStatus,
-        String headers,
-        String requestBody,
-        String responseBody
-) {
-    // ...
-}
