--- coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/response/CommonResponse.java
@@ -1,23 +1,0 @@
-package dev.be.coupon.api.common.support.response;
-
-import dev.be.coupon.api.common.support.error.ErrorMessage;
-import dev.be.coupon.api.common.support.error.ErrorType;
-
-public record CommonResponse<T>(ResultType resultType, T data, ErrorMessage errorMessage) {
-
-    public static CommonResponse<?> success() {
-        return new CommonResponse<>(ResultType.SUCCESS, null, null);
-    }
-
-    public static <T> CommonResponse<T> success(final T data) {
-        return new CommonResponse<>(ResultType.SUCCESS, data, null);
-    }
-
-    public static CommonResponse<?> error(final ErrorType errorType) {
-        return new CommonResponse<>(ResultType.ERROR, null, new ErrorMessage(errorType));
-    }
-
-    public static CommonResponse<?> error(final ErrorType error, final Object errorData) {
-        return new CommonResponse<>(ResultType.ERROR, null, new ErrorMessage(error, errorData));
-    }
-}
