+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/config/WebMvcConfig.java
@@ -1,0 +1,23 @@
+package dev.be.coupon.api.config;
+
+import dev.be.coupon.api.auth.presentation.AuthenticationPrincipalArgumentResolver;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    private final AuthenticationPrincipalArgumentResolver authenticationPrincipalArgumentResolver;
+
+    public WebMvcConfig(final AuthenticationPrincipalArgumentResolver authenticationPrincipalArgumentResolver) {
+        this.authenticationPrincipalArgumentResolver = authenticationPrincipalArgumentResolver;
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+        argumentResolvers.add(authenticationPrincipalArgumentResolver);
+    }
+}
