+++ coupon/coupon-domain/src/main/java/dev/be/coupon/domain/coupon/CouponType.java
@@ -1,0 +1,8 @@
+package dev.be.coupon.domain.coupon;
+
+/**
+ * 쿠폰 타입을 정의하는 Enum.
+ */
+public enum CouponType {
+    CHICKEN, PIZZA, BURGER
+}
