+++ support/common/build.gradle
@@ -1,0 +1,6 @@
+jar.enabled = true
+
+dependencies {
+    implementation 'org.springframework:spring-context'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
