+++ support/common/src/main/java/dev/be/coupon/common/support/error/ErrorCode.java
@@ -1,0 +1,6 @@
+package dev.be.coupon.common.support.error;
+
+public enum ErrorCode {
+    E500,
+    E400, E401, E403, E404
+}
