+++ support/common/src/main/java/dev/be/coupon/common/support/error/ErrorMessage.java
@@ -1,0 +1,33 @@
+package dev.be.coupon.common.support.error;
+
+
+public class ErrorMessage {
+
+    private final String code;
+    private final String message;
+    private final Object data;
+
+    public ErrorMessage(final ErrorType errorType) {
+        this.code = errorType.getCode().name();
+        this.message = errorType.getMessage();
+        this.data = null;
+    }
+
+    public ErrorMessage(final ErrorType errorType, final Object data) {
+        this.code = errorType.getCode().name();
+        this.message = errorType.getMessage();
+        this.data = data;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public Object getData() {
+        return data;
+    }
+}
