+++ support/common/src/main/java/dev/be/coupon/common/support/response/ResultType.java
@@ -1,0 +1,5 @@
+package dev.be.coupon.common.support.response;
+
+public enum ResultType {
+    SUCCESS, ERROR
+}
